Replace SD driver w/ upgraded experimental version
Resolves #1105
