+++ demos/sjtwo/sd_card/project.mk
@@ -1,2 +1,2 @@
-USER_TESTS += $(LIBRARY_DIR)/L2_HAL/memory/test/sd_test.cpp
+USER_TESTS += $(LIBRARY_DIR)/L2_HAL/memory/test/sd_experimental_test.cpp
 PLATFORM = lpc40xx
