--- demos/sjtwo/sd_card_experimental/project.mk
@@ -1,2 +1,0 @@
-USER_TESTS += $(LIBRARY_DIR)/L2_HAL/memory/test/sd_experimental_test.cpp
-PLATFORM = lpc40xx
