--- demos/sjtwo/sd_card_experimental/project_config.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-// Change the "#if 0" to "#if 1" below to see debug information for every
-// step of the SD card communication
-#if 1
-#define SJ2_LOG_LEVEL SJ2_LOG_LEVEL_DEBUG
-#endif
-
-#include "config.hpp"
