+++ library/L2_HAL/L2.mk
@@ -65,7 +65,6 @@ TESTS += $(LIBRARY_DIR)/L2_HAL/switches/test/button_test.cpp
 # Memory
 # ==============================================================================
 TESTS += $(LIBRARY_DIR)/L2_HAL/memory/test/sd_test.cpp
-TESTS += $(LIBRARY_DIR)/L2_HAL/memory/test/sd_experimental_test.cpp
 # ==============================================================================
 # Actuators
 # ==============================================================================
